1.2 Is permutation: paper code run tests
From the first attempt. I didn't wrote .ToCharArray() which is slow
because copying the underlying array.
Neither I used .CharAt(int i) which is from Java language. C# allows use
to use an index on the String class.

The space complexity is O(1) because a 128-int array save the count diff
between the two strings.
The time complexity is O(n) because we loop over the length of the input
strings.